# docs/core/openapi.yaml
openapi: 3.1.0
info:
  title: ecommerce-aigc-video API
  version: 2.0.0
  description: >
    V2 目标后端契约。全部路由前缀 /api，URL 无版本号，全站无鉴权。
    工作区级 agent 模块采用 propose -> approve 语义：propose 只生成待审 artifact，
    approve 才变为当前生效 artifact；下游只读取 approved/current。
    approve 不级联重置下游，shotprompt approve 也不重建 shots。
    新的分镜链路由显式创建 shot set 触发。上游变更通过 upstreamChanged 提示表达，
    不代表 stale，也不删除候选或选择。图像/视频选择写入 image_select_artifacts /
    video_select_artifacts，每个 shot 只有一个当前 selected image/video，重复选择用 UPSERT 覆盖。
    业务语义见 docs/core/interface.md，架构见 docs/core/arc_v2.md，数据模型见 docs/core/erd.md。
servers:
  - url: http://localhost:3000
    description: 本地开发（SERVER_PORT 默认 3000）
...
components:
  parameters:
...
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: 业务错误。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
...
paths:
...
  /api/test-runs/{runId}:
    delete:
      tags: [Platform]
      summary: 测试清理专用
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted: { type: integer }
        "403":
          $ref: "#/components/responses/Error"
...